Provide separate class which handles cached DingNet properties
